# Batch downstairs read and write ops