--- a/downstairs/Cargo.toml
+++ b/downstairs/Cargo.toml
@@ -33,6 +33,7 @@ tracing-subscriber = "0.2.19"
 tracing-opentelemetry = "0.14.0"
 uuid = { version = "0.8", features = [ "serde", "v4" ] }
 rusqlite = { version = "0.25" }
+hex = "0.4"
 
 [dev-dependencies]
 tempfile = "3"