--- a/downstairs/benches/downstairs.rs
+++ b/downstairs/benches/downstairs.rs
@@ -28,112 +28,135 @@ pub fn downstairs_rw_speed_benchmark(c: &mut Criterion) {
         ddef.extent_size().value as usize * ddef.extent_count() as usize;
 
     c.bench_function("region_write", |b| {
-        let mut rng = rand::thread_rng();
-        let mut buffer: Vec<u8> = Vec::with_capacity(total_size);
-        buffer.resize(total_size, 0u8);
-        rng.fill_bytes(&mut buffer);
+        b.iter_batched(
+            || {
+                let mut rng = rand::thread_rng();
+                let mut buffer: Vec<u8> = Vec::with_capacity(total_size);
+                buffer.resize(total_size, 0u8);
+                rng.fill_bytes(&mut buffer);
 
-        let mut writes: Vec<crucible_protocol::Write> =
-            Vec::with_capacity(num_blocks);
+                let mut writes: Vec<crucible_protocol::Write> =
+                    Vec::with_capacity(num_blocks);
 
-        for i in 0..num_blocks {
-            let eid: u64 = i as u64 / ddef.extent_size().value;
-            let offset: Block =
-                Block::new_512((i as u64) % ddef.extent_size().value);
+                for i in 0..num_blocks {
+                    let eid: u64 = i as u64 / ddef.extent_size().value;
+                    let offset: Block =
+                        Block::new_512((i as u64) % ddef.extent_size().value);
 
-            let data = BytesMut::from(&buffer[(i * 512)..((i + 1) * 512)]);
+                    let data = BytesMut::from(&buffer[(i * 512)..((i + 1) * 512)]);
 
-            writes.push(crucible_protocol::Write {
-                eid,
-                offset,
-                data: data.freeze(),
-                nonce: None,
-                tag: None,
-            });
-        }
+                    writes.push(crucible_protocol::Write {
+                        eid,
+                        offset,
+                        data: data.freeze(),
+                        nonce: None,
+                        tag: None,
+                    });
+                }
 
-        b.iter(|| {
-            region.region_write(&writes).unwrap();
-        })
+                writes
+            },
+            |writes| {
+                region.region_write(&writes).unwrap();
+            },
+            criterion::BatchSize::SmallInput,
+        );
     });
 
     c.bench_function("region_read", |b| {
-        let mut requests: Vec<crucible_protocol::ReadRequest> =
-            Vec::with_capacity(num_blocks);
+        b.iter_batched(
+            || {
+                let mut requests: Vec<crucible_protocol::ReadRequest> =
+                    Vec::with_capacity(num_blocks);
 
-        for i in 0..num_blocks {
-            let eid: u64 = i as u64 / ddef.extent_size().value;
-            let offset: Block =
-                Block::new_512((i as u64) % ddef.extent_size().value);
+                for i in 0..num_blocks {
+                    let eid: u64 = i as u64 / ddef.extent_size().value;
+                    let offset: Block =
+                        Block::new_512((i as u64) % ddef.extent_size().value);
 
-            requests.push(crucible_protocol::ReadRequest {
-                eid,
-                offset,
-                num_blocks: 1,
-            });
-        }
+                    requests.push(crucible_protocol::ReadRequest {
+                        eid,
+                        offset,
+                        num_blocks: 1,
+                    });
+                }
 
-        b.iter(|| {
-            let responses = region.region_read(&requests).unwrap();
-            for response in responses {
-                assert!(response.nonce.is_none());
-            }
-        })
+                requests
+            },
+            |requests| {
+                let responses = region.region_read(&requests).unwrap();
+                for response in responses {
+                    assert!(response.nonce.is_none());
+                }
+            },
+            criterion::BatchSize::SmallInput,
+        );
     });
 
     c.bench_function("region_write with nonce and tag", |b| {
-        let mut rng = rand::thread_rng();
-        let mut buffer: Vec<u8> = Vec::with_capacity(total_size);
-        buffer.resize(total_size, 0u8);
-        rng.fill_bytes(&mut buffer);
+        b.iter_batched(
+            || {
+                let mut rng = rand::thread_rng();
+                let mut buffer: Vec<u8> = Vec::with_capacity(total_size);
+                buffer.resize(total_size, 0u8);
+                rng.fill_bytes(&mut buffer);
 
-        let mut writes: Vec<crucible_protocol::Write> =
-            Vec::with_capacity(num_blocks);
+                let mut writes: Vec<crucible_protocol::Write> =
+                    Vec::with_capacity(num_blocks);
 
-        for i in 0..num_blocks {
-            let eid: u64 = i as u64 / ddef.extent_size().value;
-            let offset: Block =
-                Block::new_512((i as u64) % ddef.extent_size().value);
+                for i in 0..num_blocks {
+                    let eid: u64 = i as u64 / ddef.extent_size().value;
+                    let offset: Block =
+                        Block::new_512((i as u64) % ddef.extent_size().value);
 
-            let data = BytesMut::from(&buffer[(i * 512)..((i + 1) * 512)]);
+                    let data = BytesMut::from(&buffer[(i * 512)..((i + 1) * 512)]);
 
-            writes.push(crucible_protocol::Write {
-                eid,
-                offset,
-                data: data.freeze(),
-                nonce: Some(Vec::from(rng.gen::<[u8; 12]>())),
-                tag: Some(Vec::from(rng.gen::<[u8; 16]>())),
-            });
-        }
+                    writes.push(crucible_protocol::Write {
+                        eid,
+                        offset,
+                        data: data.freeze(),
+                        nonce: Some(Vec::from(rng.gen::<[u8; 12]>())),
+                        tag: Some(Vec::from(rng.gen::<[u8; 16]>())),
+                    });
+                }
 
-        b.iter(|| {
-            region.region_write(&writes).unwrap();
-        })
+                writes
+            },
+            |writes| {
+                region.region_write(&writes).unwrap();
+            },
+            criterion::BatchSize::SmallInput,
+        );
     });
 
     c.bench_function("region_read with nonce and tag", |b| {
-        let mut requests: Vec<crucible_protocol::ReadRequest> =
-            Vec::with_capacity(num_blocks);
+        b.iter_batched(
+            || {
+                let mut requests: Vec<crucible_protocol::ReadRequest> =
+                    Vec::with_capacity(num_blocks);
 
-        for i in 0..num_blocks {
-            let eid: u64 = i as u64 / ddef.extent_size().value;
-            let offset: Block =
-                Block::new_512((i as u64) % ddef.extent_size().value);
+                for i in 0..num_blocks {
+                    let eid: u64 = i as u64 / ddef.extent_size().value;
+                    let offset: Block =
+                        Block::new_512((i as u64) % ddef.extent_size().value);
 
-            requests.push(crucible_protocol::ReadRequest {
-                eid,
-                offset,
-                num_blocks: 1,
-            });
-        }
+                    requests.push(crucible_protocol::ReadRequest {
+                        eid,
+                        offset,
+                        num_blocks: 1,
+                    });
+                }
 
-        b.iter(|| {
-            let responses = region.region_read(&requests).unwrap();
-
-            for response in responses {
-                assert!(response.nonce.is_some());
-            }
-        })
+                requests
+            },
+            |requests| {
+                let responses = region.region_read(&requests).unwrap();
+                for response in responses {
+                    assert!(response.nonce.is_some());
+                }
+            },
+            criterion::BatchSize::SmallInput,
+        );
     });
 }