--- a/downstairs/src/lib.rs
+++ b/downstairs/src/lib.rs
@@ -1,0 +1,11 @@
+// Copyright 2021 Oxide Computer Company
+use anyhow::{bail, Result};
+use std::collections::HashMap;
+use std::path::PathBuf;
+
+use crucible_common::Block;
+use crucible_protocol::*;
+
+mod dump;
+mod region;
+pub use region::Region;