--- a/downstairs/src/lib.rs
+++ b/downstairs/src/lib.rs
@@ -1,4 +1,11 @@
 // Copyright 2021 Oxide Computer Company
+
+/*
+ * Note: This lib.rs was created to facilitate criterion benchmarks, which
+ * cannot benchmark binary crates. Including a minimal number of uses here
+ * allows us to write those microbenchmarks.
+ */
+
 use anyhow::{bail, Result};
 use std::collections::HashMap;
 use std::path::PathBuf;