--- a/downstairs/src/main.rs
+++ b/downstairs/src/main.rs
@@ -33,6 +33,7 @@ use uuid::Uuid;
 mod dump;
 mod region;
 mod stats;
+
 use dump::dump_region;
 use region::Region;
 use stats::*;