--- a/downstairs/src/region.rs
+++ b/downstairs/src/region.rs
@@ -787,8 +787,8 @@ impl Region {
          * together.
          *
          * Note: Have to maintain order with reads! The Upstairs expects read
-         * responses to be in the same order as read requests, so we can't use
-         * a hashmap in the same way that batching writes can.
+         * responses to be in the same order as read requests, so we can't
+         * use a hashmap in the same way that batching writes can.
          */
         let mut eid: Option<u64> = None;
         let mut batched_reads: Vec<&crucible_protocol::ReadRequest> =