Use criterion to benchmark the downstairs functions region_read and region_write. This required adding a downstairs/src/lib.rs to the project because Criterion cannot use a binary crate as a dependency.

Also added test_big_write was also added to fully validate region_write and region_read.

The main performance gain of this PR is from sending all values in a giant INSERT OR REPLACE call in set_encryption_context, followed by a secondary performance improvement of reducing the number of lock operations on `region_*` by batching calls to the extent read and write functions.

Criterion results show a large improvement to the region_write micro benchmarks:

Before:
```
Created new region file "/tmp/.tmpPeXi5o/region.json"
Benchmarking region_write: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.9s, or reduce sample count to 80.
region_write            time:   [49.274 ms 49.387 ms 49.505 ms]                         
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

region_read             time:   [39.671 ms 39.794 ms 39.943 ms]                        
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe

Benchmarking region_write with nonce and tag: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 547.3s, or reduce sample count to 10.
region_write with nonce and tag                                                                            
	                time:   [5.0685 s 5.1573 s 5.2230 s]
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) low severe
  1 (1.00%) high mild

region_read with nonce and tag                                                                            
	                time:   [43.019 ms 43.183 ms 43.362 ms]
Found 11 outliers among 100 measurements (11.00%)
  7 (7.00%) high mild
  4 (4.00%) high severe
```

After:
```
Created new region file "/tmp/.tmpwWNFcL/region.json"
region_write            time:   [11.659 ms 11.719 ms 11.784 ms]                         
	                change: [-76.404% -76.271% -76.131%] (p = 0.00 < 0.05)
	                Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe

region_read             time:   [41.086 ms 41.237 ms 41.404 ms]                        
	                change: [+3.0621% +3.6272% +4.1825%] (p = 0.00 < 0.05)
	                Performance has regressed.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

Benchmarking region_write with nonce and tag: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 57.1s, or reduce sample count to 10.
region_write with nonce and tag                                                                            
	                time:   [539.00 ms 542.17 ms 545.58 ms]
	                change: [-89.637% -89.487% -89.295%] (p = 0.00 < 0.05)
	                Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  8 (8.00%) high mild
  3 (3.00%) high severe

region_read with nonce and tag                                                                            
	                time:   [43.466 ms 43.515 ms 43.568 ms]
	                change: [+0.3467% +0.7711% +1.1765%] (p = 0.00 < 0.05)
	                Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
```