@jmpesp 2021-12-17
> Could we put instructions on how to run this in the README.md?
it's just cargo bench right?
Is there a way to run just the bench specific tests without all the noise that is generated when I run cargo bench. By noise, I mean all the ignored tests.

Done in 364bff6