@jmpesp 2021-12-17
>I'm also curious about a few things, and perhaps first is a general note of caution about taking action on benchmark results too soon. I don't want to optimize too much before we have gotten a better idea on what our block size and blocks per extent is going to be.
>
> Building the tools to benchmark, that is very important, and the benchmarks here are good to have. But think first getting benchmarks from the guest and seeing where the rough spots are there might give us a better idea of where to focus. There could be other low hanging fruit which we should pick first.

I agree, but in this case (plus the follow up PR) there was some *really* low hanging fruit that should be picked immediately :)